Module Migemocaml

Migemocaml is the one of the implementation of migemo that is written Ruby originally.

module Migemo : sig ... end

Migemo module provides core function to query string with migemo algorithm.

module Regexp_spec : sig ... end

Regexp_spec define module interface and some modules to generate regexp from Migemo module.

module Dict_tree : sig ... end

Dict_tree module provides tree structure for migemo dictionary.