Module Migemocaml__Converter_character

Converter_character module provides basement function to convert a character to other character. This module will be able to use conversion between romaji and other character, but should use Converter_romaji to get valid result.

module Core : sig ... end

include Core

type t = Migemocaml.Dict_tree.t

val convert_character : string -> Migemocaml.Dict_tree.t -> string * int

convert a character from string with tree

val convert : string:string -> Migemocaml.Dict_tree.t -> string

Convert string to other characters with tree. This function read string as romaji.